type_store: fix type_store_lookup_alias behavior
It was an outlier among type_store_lookup_* functions for requiring the
caller to manually construct the representation of the type that it
wished to put into the type store, while all the other functions do this
by themselves. The distinction existed primarily because alias creation
is a convoluted process where the information required to constuct the
whole type has annoying interdependencies.

The new interface does not solve this completely - unlike other similar
functions, this one has to permit passing NULL as secondary type into
it, and as a result, needs a separate flags parameter. I expect these to
both be resolved once we determine what kind of behavior we actually
want from type aliases.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>