gen: fix iterator variable intialization
gen_copy_aligned is not really meant for general consumption, and leads
to worse output in cases when the iterator variable is a primitive type.
Changing this unmasked an issue with the type field of gptr gen_value,
which is also fixed now. A test is added for a case that helped me
identify this.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>