gen: simplify branch generation
Introduce gen_expr_branch that is basically an enhanced combination of
gen_expr_with and branch_copyresult. The check from branch_copyresult is
reduced to the usual non-void non-never check on one type only, this is
enough because check guarantees that the branch and merged types will
either match, or the branch type will be never.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>