scan: make incomplete_decl just a field in scope_object
It doesn't make sense to treat O_SCAN objects differently than everything
else. I'm not sure it was ever a good idea, but it certainly isn't a
good idea now, because it just makes the interface harder to use
correctly.

Rename struct incomplete_declaration to incomplete_decl. It's shorter,
unambiuous, and this commit touches nearly every line that uses it
anyway.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>