gen: fix a regression introduced in 13a6fc7df4c4
alloc(obj, cap) called memcpy with too big size parameter when `obj` was
a slice, causing copying of unrelated memory next to initialzier
object's memory location. This is a stupid thing to do, but strictly
speaking, it's was actually almost legal and mostly harmless, since the
copied memory was inaccessible from within hare without pointer hacks.

The only thing this actually broke was allocation of empty slices (see
added test), because when the initializer slice points to null, trying
to copy anything out of it is a bad idea.