Ensure that unnamed parameters end up in the ABI.
Lorenz pointed out that

	fn foo(_: str, _: int, x: int) int = 2 * x;

was returning 20 when called like foo("unused", 10, 5), when instead it
should return 10. I found that is the result of ddfbdd8d, in which I
allowed unnamed parameters in functions with a body, as dictated by the
spec. The above function compiles to $foo(w %x), skipping the unused
parameters entirely. QBE doesn't mind if you call functions with the
wrong number of arguments, because it has all the ABI information in the
call instruction.

I had tested it using fn foo(_: str, x: int) int = 2 * x, which only
worked because the unused str would be passed on the stack. It was only
when you added the extra register-passed argument that it misbehaves.

To ensure the parameters end up in the ABI, we need to name them. Here I
just name them all "_". QBE doesn't seem to mind duplicated names. If that
breaks at some point then we can give them generated unique names.

Signed-off-by: Joe Finney <me@spxtr.net>