Give unions a default value.
If any of their fields have a default value, so do they. Until
@default exists, the default value is just zero for all types. If/when
@default happens, this will need to be revisited.

A good alternative might be making it so that unions do not need to be
initialized at all, or they can only be initialized with whatever
unsafe initializer syntax we come up with. Unions are fundamentally
unsafe.

Signed-off-by: Joe Finney <me@spxtr.net>